add a trivial outer constructor for Binomial
Before, Binomial could only be called with arguments of type
(Vector{T}, Int64, Int64). Now it can be called with
(AbstractVector{T}, Integer, Integer), with the appropriate conversions
being done automatically when constructing the concrete type.

This fixes #93 and also makes it possible to call:

```
subsets(1:3, 2)
```

instead of

```
subsets(collect(1:3), 2)
```