Add cached iterator
For iterators that are expensive to compute and cheap to store, you might want to cache the result so the next time you iterate it just returns items from the cache. In a way this is kind of like lazy sequences from Clojure/Lazy.jl

This is missing documentation and tests but would like to gauge interest before committing time to that.